# lib/api-zod/src/generated/types/chartAnalysisRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChartAnalysisRequestVisualBias } from './chartAnalysisRequestVisualBias';

export interface ChartAnalysisRequest {
  asset: string;
  filename: string;
  /**
     * @minimum 1
     * @maximum 10000
     */
  imageWidth: number;
  /**
     * @minimum 1
     * @maximum 10000
     */
  imageHeight: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  bullishCandleRatio: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  bearishCandleRatio: number;
  /**
     * @minimum -1
     * @maximum 1
     */
  trendScore: number;
  /**
     * @minimum 1
     * @maximum 1000
     */
  sampledColumns: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  visualConfidence: number;
  visualBias: ChartAnalysisRequestVisualBias;
  /**
     * Normalized vertical position of the highest detected chart region; 0 is the top of the plot.
     * @minimum 0
     * @maximum 1
     */
  visibleHighPosition: number;
  /**
     * Normalized vertical position of the lowest detected chart region; 1 is the bottom of the plot.
     * @minimum 0
     * @maximum 1
     */
  visibleLowPosition: number;
  /**
     * Normalized vertical position of the latest detected chart region.
     * @minimum 0
     * @maximum 1
     */
  lastPricePosition: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  priceRangeConfidence: number;
  /**
     * Optional readable price label at the top of the chart plot.
     * @minimum 0
     */
  priceScaleTop?: number;
  /**
     * Optional readable price label at the bottom of the chart plot.
     * @minimum 0
     */
  priceScaleBottom?: number;
  /**
     * @maxItems 8
     * @items.maxLength 180
     */
  signals: string[];
}