# lib/api-zod/src/generated/types/chartPriceRange.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChartPriceRange {
  lower: number;
  upper: number;
  midpoint: number;
  currentPrice: number;
  expectedMovePct: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  source: string;
  note: string;
}