# lib/api-zod/src/generated/types/featureExplanation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeatureExplanationDirection } from './featureExplanationDirection';

export interface FeatureExplanation {
  feature: string;
  label: string;
  value: number;
  contribution: number;
  direction: FeatureExplanationDirection;
}