# lib/api-zod/src/generated/types/modelComparison.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ModelComparison {
  model: string;
  /** @nullable */
  accuracy: number | null;
  /** @nullable */
  f1: number | null;
}