# lib/api-zod/src/generated/types/modelInfo.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ModelInfoStatus } from './modelInfoStatus';

export interface ModelInfo {
  status: ModelInfoStatus;
  modelName: string;
  version: string;
  featureCount: number;
  features: string[];
  metadataAvailable: boolean;
  /** @nullable */
  modelPath?: string | null;
  message?: string;
}