# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionDirection } from './predictionDirection';
import type { PredictionModelStatus } from './predictionModelStatus';

export interface Prediction {
  asset: string;
  direction: PredictionDirection;
  /**
     * @minimum 0
     * @maximum 1
     */
  probability: number;
  horizon: string;
  timestamp: Date;
  modelStatus: PredictionModelStatus;
  disclaimer?: string;
}