# lib/api-zod/src/generated/types/realtimeAnalysisResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChartPriceRange } from './chartPriceRange';
import type { ChartScenario } from './chartScenario';
import type { MarketCandle } from './marketCandle';
import type { Prediction } from './prediction';
import type { RealtimeAnalysisResponseCombinedDirection } from './realtimeAnalysisResponseCombinedDirection';
import type { RealtimeAnalysisResponseProviderStatus } from './realtimeAnalysisResponseProviderStatus';
import type { TechnicalIndicators } from './technicalIndicators';

export interface RealtimeAnalysisResponse {
  provider: string;
  providerStatus: RealtimeAnalysisResponseProviderStatus;
  providerMessage: string;
  symbol: string;
  timeframe: string;
  updatedAt: Date;
  currentPrice: number;
  candles: MarketCandle[];
  indicators: TechnicalIndicators;
  model: Prediction;
  combinedDirection: RealtimeAnalysisResponseCombinedDirection;
  priceRange: ChartPriceRange;
  /**
     * @minItems 3
     * @maxItems 3
     */
  scenarios: ChartScenario[];
  disclaimer: string;
}