# lib/api-zod/src/generated/types/technicalIndicators.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TechnicalIndicators {
  ema9: number;
  ema21: number;
  sma50: number;
  rsi: number;
  macd: number;
  bollingerUpper: number;
  bollingerLower: number;
  atr: number;
}